Avances

+++ app/Http/Controllers/EmpleadoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Empleado;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class EmpleadoController extends Controller
 {
@@ -39,6 +40,23 @@ class EmpleadoController extends Controller
     public function store(Request $request)
     {
         //
+        $campos=[
+            'Nombre'=>'required|string|max:100',
+            'ApellidoPaterno'=>'required|string|max:100',
+            'ApellidoPaterno'=>'required|string|max:100',
+            'NumeroDocumento'=>'required|string|max:12',
+            'Correo'=>'required|email',
+            'Direccion'=>'required|string|max:100',
+            'Telefono'=>'required|string|max:11',
+            'Genero'=>'required|string|max:20',
+            'Foto'=>'required|max:10000|mimes:jpeg,jpg,png',
+        ];
+        $mensaje=[
+            'required'=>'El :attribute es requerido',
+            'Foto.required'=>'La foto es requerida',
+        ];
+
+        $this->validate($request, $campos, $mensaje);
         //$datosEmpleados = request()->all();
         $datosEmpleados = request()->except('_token');
         if ($request->hasFile('Foto')) {
@@ -46,7 +64,8 @@ class EmpleadoController extends Controller
             # code...
         }
         Empleado::insert($datosEmpleados);
-        return response()->json($datosEmpleados);
+        //return response()->json($datosEmpleados);
+        return redirect('empleado')->with('mensaje','Empleado agregado con exito!.');
     }
 
     /**
@@ -66,9 +85,11 @@ class EmpleadoController extends Controller
      * @param  \App\Models\Empleado  $empleado
      * @return \Illuminate\Http\Response
      */
-    public function edit(Empleado $empleado)
+    public function edit($id)
     {
         //
+        $empleado = Empleado::findOrFail($id);
+        return view('empleado.edit', compact('empleado'));
     }
 
     /**
@@ -78,9 +99,46 @@ class EmpleadoController extends Controller
      * @param  \App\Models\Empleado  $empleado
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Empleado $empleado)
+    public function update(Request $request, $id)
     {
+        $campos=[
+            'Nombre'=>'required|string|max:100',
+            'ApellidoPaterno'=>'required|string|max:100',
+            'ApellidoPaterno'=>'required|string|max:100',
+            'NumeroDocumento'=>'required|string|max:12',
+            'Correo'=>'required|email',
+            'Direccion'=>'required|string|max:100',
+            'Telefono'=>'required|string|max:11',
+            'Genero'=>'required|string|max:20',
+
+        ];
+        $mensaje=[
+            'required'=>'El :attribute es requerido',
+
+        ];
+
+        if ($request->hasFile('Foto')) {
+            $campos=['Foto'=>'required|max:10000|mimes:jpeg,jpg,png'];
+            $mensaje=['Foto.required'=>'La foto es requerida'];
+        }
+
+        $this->validate($request, $campos, $mensaje);
         //
+        $datosEmpleados = request()->except(['_token','_method']);
+
+        if ($request->hasFile('Foto')) {
+            $empleado = Empleado::findOrFail($id);
+            Storage::delete('public/'.$empleado->Foto);
+            $datosEmpleados['Foto']=$request->file('Foto')->store('uploads','public');
+            # code...
+        }
+
+        Empleado::where('id','=',$id)->update($datosEmpleados);
+
+        $empleado = Empleado::findOrFail($id);
+        //return view('empleado.edit', compact('empleado'));
+
+        return redirect('empleado')->with('mensaje','Empleado editado correctamente!.');
     }
 
     /**
@@ -92,7 +150,12 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+        if (Storage::delete('public/'.$empleado->Foto)) {
+            Empleado::destroy($id);
+            # code...
+        }
+
+        return redirect('empleado')->with('mensaje','Empleado borrado!.');
     }
 }
